// src/generated/types.tsx
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Aggregation = {
  __typename?: 'Aggregation';
  /** Attribute code of the aggregation group. */
  attribute_code: Scalars['String'];
  /** The number of options in the aggregation group. */
  count?: Maybe<Scalars['Int']>;
  /** The aggregation display name. */
  label?: Maybe<Scalars['String']>;
  /** Array of options for the aggregation. */
  options?: Maybe<Array<Maybe<AggregationOption>>>;
  /** The relative position of the attribute in a layered navigation block */
  position?: Maybe<Scalars['Int']>;
};

export type AggregationOption = AggregationOptionInterface & {
  __typename?: 'AggregationOption';
  /** The number of items that match the aggregation option. */
  count?: Maybe<Scalars['Int']>;
  /** Aggregation option display label. */
  label?: Maybe<Scalars['String']>;
  /** The internal ID that represents the value of the option. */
  value: Scalars['String'];
};
// ...
export type Breadcrumb = {
  __typename?: 'Breadcrumb';
  /**
   * Category ID.
   * @deprecated Use the `category_uid` argument instead.
   */
  category_id?: Maybe<Scalars['Int']>;
  /** Category level. */
  category_level?: Maybe<Scalars['Int']>;
  /** Category name. */
  category_name?: Maybe<Scalars['String']>;
  /** The unique ID for a `Breadcrumb` object. */
  category_uid: Scalars['ID'];
  /** Category URL key. */
  category_url_key?: Maybe<Scalars['String']>;
  /** Category URL path. */
  category_url_path?: Maybe<Scalars['String']>;
};
// ...
export type CategoryFilterInput = {
  /** Filter by the unique category ID for a `CategoryInterface` object. */
  category_uid?: InputMaybe<FilterEqualTypeInput>;
  /** Deprecated: use 'category_uid' to filter uniquely identifiers of categories. */
  ids?: InputMaybe<FilterEqualTypeInput>;
  /** Filter by the display name of the category. */
  name?: InputMaybe<FilterMatchTypeInput>;
  /** Filter by the unique parent category ID for a `CategoryInterface` object. */
  parent_category_uid?: InputMaybe<FilterEqualTypeInput>;
  /** Filter by the unique parent category ID for a `CategoryInterface` object. */
  parent_id?: InputMaybe<FilterEqualTypeInput>;
  /** Filter by the part of the URL that identifies the category. */
  url_key?: InputMaybe<FilterEqualTypeInput>;
  /** Filter by the URL path for the category. */
  url_path?: InputMaybe<FilterEqualTypeInput>;
};
// ...
export type Products = {
  __typename?: 'Products';
  /** Layered navigation aggregations. */
  aggregations?: Maybe<Array<Maybe<Aggregation>>>;
  /**
   * Layered navigation filters array.
   * @deprecated Use aggregations instead
   */
  filters?: Maybe<Array<Maybe<LayerFilter>>>;
  /** An array of products that match the specified search criteria. */
  items?: Maybe<Array<Maybe<ProductInterface>>>;
  /** An object that includes the page_info and currentPage values specified in the query. */
  page_info?: Maybe<SearchResultPageInfo>;
  /** An object that includes the default sort field and all available sort fields. */
  sort_fields?: Maybe<SortFields>;
  /** The number of products that are marked as visible. By default, in complex products, parent products are visible, but their child products are not. */
  total_count?: Maybe<Scalars['Int']>;
};

export type Query = {
  __typename?: 'Query';
  /** Get a list of available store views and their config information. */
  availableStores?: Maybe<Array<Maybe<StoreConfig>>>;
  /** Returns information about shopping cart */
  cart?: Maybe<Cart>;
  categories?: Maybe<CategoryResult>;
  /**
   * The category query searches for categories that match the criteria specified in the search and filter attributes.
   * @deprecated Use 'categoryList' query instead of 'category' query
   */
  category?: Maybe<CategoryTree>;
  /** Returns an array of categories based on the specified filters. */
  categoryList?: Maybe<Array<Maybe<CategoryTree>>>;
  /** Retrieves an array of configuration data for the chat widget. */
  chatData?: Maybe<ChatData>;
  /** The Checkout Agreements information */
  checkoutAgreements?: Maybe<Array<Maybe<CheckoutAgreement>>>;
  /** The CMS block query returns information about CMS blocks */
  cmsBlocks?: Maybe<CmsBlocks>;
  /** The CMS page query returns information about a CMS page */
  cmsPage?: Maybe<CmsPage>;
  /** Return products that have been added to the specified compare list */
  compareList?: Maybe<CompareList>;
  /** The countries query provides information for all countries. */
  countries?: Maybe<Array<Maybe<Country>>>;
  /** The countries query provides information for a single country. */
  country?: Maybe<Country>;
  /** The currency query returns information about store currency. */
  currency?: Maybe<Currency>;
  /** The customAttributeMetadata query returns the attribute type, given an attribute code and entity type */
  customAttributeMetadata?: Maybe<CustomAttributeMetadata>;
  /** The customer query returns information about a customer account */
  customer?: Maybe<Customer>;
  /** Returns information about the customer shopping cart */
  customerCart: Cart;
  /** The query returns the contents of a customer's downloadable products */
  customerDownloadableProducts?: Maybe<CustomerDownloadableProducts>;
  /** @deprecated Use orders from customer instead */
  customerOrders?: Maybe<CustomerOrders>;
  /** Return a list of customer payment tokens */
  customerPaymentTokens?: Maybe<CustomerPaymentTokens>;
  /** Returns status of Easy Email Capture for Checkout. */
  emailCaptureCheckout?: Maybe<IsConfigSettingEnabledOutput>;
  /** Returns status of Easy Email Capture for Newsletter. */
  emailCaptureNewsletter?: Maybe<IsConfigSettingEnabledOutput>;
  /** Retrieve secure PayPal url for Payments Pro Hosted Solution transaction. */
  getHostedProUrl?: Maybe<HostedProUrl>;
  /** Retrieve payment credentials for transaction. Use this query for Payflow Link and Payments Advanced payment methods. */
  getPayflowLinkToken?: Maybe<PayflowLinkToken>;
  isEmailAvailable?: Maybe<IsEmailAvailableOutput>;
  /** Retrieves information about an order by order id. */
  orderData?: Maybe<Order>;
  /** The pickup locations query searches for locations that match the search request requirements. */
  pickupLocations?: Maybe<PickupLocations>;
  /** Retrieves metadata required by clients to render the Reviews section. */
  productReviewRatingsMetadata: ProductReviewRatingsMetadata;
  /** The products query searches for products that match the criteria specified in the search and filter attributes. */
  products?: Maybe<Products>;
  /** Return the full details for a specified product, category, or CMS page given the specified url_key, appended by the url_suffix, if one exists */
  route?: Maybe<RoutableInterface>;
  /** The store config query */
  storeConfig?: Maybe<StoreConfig>;
  todos?: Maybe<Todos>;
  /** Retrieves an array of configuration data for different types of tracking. */
  trackingData?: Maybe<TrackingData>;
  /**
   * The urlResolver query returns the relative URL for a specified product, category or CMS page, using as input a url_key appended by the url_suffix, if one exists
   * @deprecated Use the 'route' query instead
   */
  urlResolver?: Maybe<EntityUrl>;
  /**
   * The wishlist query returns the contents of a customer's wish list
   * @deprecated Moved under `Customer` `wishlist`
   */
  wishlist?: Maybe<WishlistOutput>;
};
// ...
export type AggregationFragment = {
  __typename?: 'Aggregation';
  attribute_code: string;
  count?: number | null | undefined;
  label?: string | null | undefined;
  position?: number | null | undefined;
  options?:
    | Array<
        | {
            __typename?: 'AggregationOption';
            count?: number | null | undefined;
            label?: string | null | undefined;
            value: string;
          }
        | null
        | undefined
      >
    | null
    | undefined;
};

export type CategoryTreeFragment = {
  __typename?: 'CategoryTree';
  uid: string;
  url_key?: string | null | undefined;
  url_path?: string | null | undefined;
  name?: string | null | undefined;
  id: string;
  breadcrumbs?:
    | Array<
        | {
            __typename?: 'Breadcrumb';
            category_url_path?: string | null | undefined;
            category_name?: string | null | undefined;
            id?: number | null | undefined;
          }
        | null
        | undefined
      >
    | null
    | undefined;
  children?:
    | Array<
        | {
            __typename?: 'CategoryTree';
            uid: string;
            url_key?: string | null | undefined;
            url_path?: string | null | undefined;
            name?: string | null | undefined;
            id: string;
          }
        | null
        | undefined
      >
    | null
    | undefined;
};
// ...
export type CategoryQueryVariables = Exact<{
  filters?: InputMaybe<CategoryFilterInput>;
}>;

export type CategoryQuery = {
  __typename?: 'Query';
  categoryList?:
    | Array<
        | {
            __typename?: 'CategoryTree';
            uid: string;
            url_key?: string | null | undefined;
            url_path?: string | null | undefined;
            name?: string | null | undefined;
            id: string;
            breadcrumbs?:
              | Array<
                  | {
                      __typename?: 'Breadcrumb';
                      category_url_path?: string | null | undefined;
                      category_name?: string | null | undefined;
                      id?: number | null | undefined;
                    }
                  | null
                  | undefined
                >
              | null
              | undefined;
            children?:
              | Array<
                  | {
                      __typename?: 'CategoryTree';
                      uid: string;
                      url_key?: string | null | undefined;
                      url_path?: string | null | undefined;
                      name?: string | null | undefined;
                      id: string;
                    }
                  | null
                  | undefined
                >
              | null
              | undefined;
          }
        | null
        | undefined
      >
    | null
    | undefined;
};
// ...
export type ProductsQueryVariables = Exact<{
  search?: InputMaybe<Scalars['String']>;
  filters?: InputMaybe<ProductAttributeFilterInput>;
  pageSize?: InputMaybe<Scalars['Int']>;
  page?: InputMaybe<Scalars['Int']>;
}>;
// ...
export const CategoryTreeFragmentDoc = gql`
  fragment categoryTree on CategoryTree {
    id: uid
    uid
    url_key
    url_path
    name
    breadcrumbs {
      id: category_id
      category_url_path
      category_name
    }
    children {
      id: uid
      uid
      url_key
      url_path
      name
    }
  }
`;
// ...
export const ProductsFragmentDoc = gql`
  fragment products on Products {
    aggregations {
      ...aggregation
    }
    page_info {
      current_page
      total_pages
    }
    items {
      ...productBundle
      ...productConfigurable
      ...productDownloadable
      ...productGrouped
      ...productSimple
      ...productVirtual
    }
  }
  ${AggregationFragmentDoc}
  ${ProductBundleFragmentDoc}
  ${ProductConfigurableFragmentDoc}
  ${ProductDownloadableFragmentDoc}
  ${ProductGroupedFragmentDoc}
  ${ProductSimpleFragmentDoc}
  ${ProductVirtualFragmentDoc}
`;
// ...
export const CategoryDocument = gql`
  query Category($filters: CategoryFilterInput) {
    categoryList(filters: $filters) {
      ...categoryTree
    }
  }
  ${CategoryTreeFragmentDoc}
`;
// ...
export const ProductsDocument = gql`
  query Products(
    $search: String
    $filters: ProductAttributeFilterInput
    $pageSize: Int = 9
    $page: Int = 1
  ) {
    products(
      search: $search
      filter: $filters
      pageSize: $pageSize
      currentPage: $page
    ) {
      ...products
    }
  }
  ${ProductsFragmentDoc}
`;